backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
accelerate==1.1.1
backoff==2.2.1
boto3==1.35.71
botocore==1.35.71
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gigachain==0.2.16
gigachain-community==0.2.16.post3
gigachain-core==0.2.38.post2
gigachain-text-splitters==0.2.4
gigachat==0.1.36
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langfuse==2.54.1
langsmith==0.1.146
markupsafe==2.1.5
marshmallow==3.23.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
pydub==0.25.1
pypdf==5.1.0
python-multipart==0.0.12
chardet==5.2.0
ruff==0.8.0
semantic-version==2.10.0
tomlkit==0.12.0
unstructured==0.16.8
networkx==3.4.2
jq==1.8.0
numpy==1.26.4
openai==1.55.1
orjson==3.10.12
packaging==24.2
propcache==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
setuptools==75.6.0
sentence-transformers==3.3.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
wrapt==1.17.0
yarl==1.18.0
onnxruntime==1.20.1
python_docx==1.1.2
youtube_transcript_api==0.6.3
selenium==4.27.1
webdriver_manager==4.0.2
fake_useragent==1.5.1
beautifulsoup4==4.12.3
faiss-cpu==1.9.0.post1
ffmpy==0.4.0
